+++ RayTraceView.h
@@ -56,6 +56,7 @@ private:
 	POINT		m_AltStart;
 	void GetVectorFromPoint(sp& k, sp& l, int px, int py);
 
+
 	//Original image width and height
 	int imageW, imageH;
 
@@ -63,14 +64,14 @@ private:
 	int crunch;		// no use
 
 	// Starting position and scale
-	double xOff;	// no use
-	double yOff;	// no use
-	double scale;
+	float xOff;	// no use
+	float yOff;	// no use
+	float scale;
 
 	// Starting stationary position and scale motion
-	double xdOff;	// no use
-	double ydOff;	// no use
-	double dscale;	// no use
+	float xdOff;	// no use
+	float ydOff;	// no use
+	float dscale;	// no use
 
 	// Starting animation frame and anti-aliasing pass 
 	int animationFrame;
