+++ node.h
@@ -17,10 +17,10 @@ struct	matrix;
 
 struct fsize
 {
-	double	top;
-	double	left;
-	double	bottom;
-	double	right;
+	float	top;
+	float	left;
+	float	bottom;
+	float	right;
 };
 #define PERSPECTIVE_RATIO 0.1
 
@@ -38,13 +38,13 @@ protected:
 	matrix		m_Pivot;	// マニュピレータの中心点
 	matrix		m_Matrix;
 
-	double		m_Reflect ;		// 反射率
-	double		m_Through ;  	// 透過率
-	double		m_Refractive ;	// 屈折率
+	float		m_Reflect ;		// 反射率
+	float		m_Through ;  	// 透過率
+	float		m_Refractive ;	// 屈折率
 
 	struct tagBoundary {
 		sp		Center;	// 中心
-		double	Radius;	// 半径
+		float	Radius;	// 半径
 		tagBoundary() : Center(), Radius(1) {}
 	} m_Boundary;	// 境界
 
@@ -79,16 +79,16 @@ public:
 
 			sp GetColor(const sp& K, const sp& L, int nest = 0) const;
 			BOOL GetInfo2(const sp& K, const sp& L, Info& info) const;
-			sp GetPixel(double x, double y) const;
+			sp GetPixel(float x, float y) const;
 	virtual	BOOL GetInfo(const sp& K, const sp& L, Info& info) const = 0;
 	virtual	BOOL IsInside(const sp& L) const = 0;
 	
-	void Move(eAxis axis, double d);
+	void Move(eAxis axis, float d);
 	void Move(POINT d);
-	void Rotate(eAxis axis, double d);
+	void Rotate(eAxis axis, float d);
 	void Rotate(POINT d);
-	void Scale(eAxis axis, double d);
-	void MovePivot(eAxis axis, double d);
+	void Scale(eAxis axis, float d);
+	void MovePivot(eAxis axis, float d);
 
 	virtual void Draw_Outline(CDC* pDC, CRayTraceView& raytraceview, const matrix& Matrix) const;
 	virtual bool SetManipulatorAxis(CRayTraceView& rtv, CPoint point, const matrix& m) const;
