+++ rotate_z.cpp
@@ -1,8 +1,8 @@
 #include "stdafx.h"
 
-rotate_z::rotate_z( double angle ) : matrix( 4, 4 )
+rotate_z::rotate_z( float angle ) : matrix( 4, 4 )
 {
-	double	th = angle / 180.0 * M_PI;
+	float	th = angle / 180.0 * M_PI;
 
 	m_data[0][0] = m_data[1][1] = cos(th);
 	m_data[1][0] = sin(th);
